Add CGI-backed comments section.
The CGI script for form submission is added to src/cgi-bin directory.
Currently, I use fcgiwrap and nginx on my server to run the script.
Forms are submitted to a sqlite3 database if they pass the basic
puzzle test. The puzzle is solved by comparing the submitted value
to the value of an environment value. Let's see how goot it performs.

Moreover, the build script and Makefile are updated accordingly.